requirements.txt
asgiref==1.1.2
attrs==17.2.0
autobahn==17.7.1
Automat==0.6.0
channels==1.1.6
constantly==15.1.0
daphne==1.3.0
Django==1.11.4
django-object-actions==0.10.0
django-password-reset==1.0
django-rest-framework==0.1.0
djangorestframework==3.6.3
djangorestframework-jwt==1.11.0
hyperlink==17.3.0
incremental==17.5.0
olefile==0.44
Pillow==4.2.1
psycopg2==2.7.3
PyJWT==1.5.2
pytz==2017.2
six==1.10.0
Twisted==17.5.0
txaio==2.8.1
zope.interface==4.4.2
